Refs codership/galera-bugs#165 Added encryption support in free()/discard()
This commit adds encryption support in free(), discard() and
seqno_assign() methods. In particular, it makes seqno_assign() to mark
plaintext buffer as changed even if it was previously flushed to disk
so that another flush will be triggered on free(). It also adds
BufferHeader copy to plaintext context, so that iterations over
containers don't require recovery of the plaintext if it happens to be
flushed.