codership/galera-bugs#714 handshake failed with duplicate uuid
The handshake failure with 'duplicate uuid' message was occasionally
seen when wsrep TLS service was enabled. The reason was that the
address scheme was not always set to ssl, which led to duplicate
address entries in gmcast connection tracking with mixed address
schemes. This confused the duplicate uuid detection logic, causing
occasional false positives during server restarts.

As a fix, always assign ssl scheme when wsrep TLS service is
enabled, except if dynamic protocol is configured.