MENT-1333: fix data type for some fields in P_S for Galera
Currently, the values of wsrep_apply_window and wsrep_commit_window
are shown differently in galera_group_members and global_status tables.

To immediately fix this specific problem, it is enough to change
the data types in the structure that is passed from Galera via
the Performance Schema-specific API. However, this change will
make the Galera library incompatible with older server versions.
Therefore, we need to change the names of the entry points in it.
In addition, in order to avoid similar problems in the future and
ensure compatibility between different versions of the library
and the server, it is preferable to add a field with a version
number to the data structure itself (for server-side verification
and the possibility in the future to make a union between several
structures, if another change to the data structure will be made).
We also need to move the memory allocation for this data structure
from the server to the Galera library itself, since the server may
not know the length of the structure for this particular version
of the library. The given commit makes these changes. These changes
can be tested without adding a new test for mtr.