Fix writeset waiting for terminate_trx()
Function terminate_trx() sends a writeset containing a commit /
rollback fragment to terminate a prepared transaction. The function
uses repl->send() to send the fragment, followed by sync_wait() to
wait for the delivery of the fragment. It turns out that sync_wait()
may return before the fragment is delivered.
This patch therefore introduces a new mchanims to wait for a writeset
from a given transaction, to be used instead of the sync wait
operation.