codership/galera-bugs#955 Asio acceptor stops accepting connections
Prior to this fix, if an error happened during streaming engine
server handshake, the acceptor stopped accepting new connections.
Also the handshake for the newly accepted connection had to
complete before a new connection could be accepted.

A new testcase test_accept_after_server_handshake_error is
added to verify that accepting continues after handshake error.
In order to allow error injection for server side handshake
testing, `AsioAcceptor` interface is changed to take
`AsioSocketHandler` as an argument.

The actual fix is to notify acceptor_handler immediately
after a new socket has been accepted. The responsibility to
keep the newly accepted socket alive during the handshake is
shifted to upper layer. It is also enforced that trying to
read or write to the socket before the handshake is over produces
an error. The upper layer must wait for notification delivered
through `AsioSocketHandler::connect_handler()` before attempting
to read or write.

Logging cleanup: Messages that don't require action from an
administrator are removed or changed to information level.