Create self signed certificate chain for testing
The previous code to create certificate chains for testing
was not correct, and was not caught because of missing check
in testcase test_ssl_certificate_chaing.

Fix by generating self signed certificate chain for testing
purposes. Add check in test_ssl_certificate_chain to verify
that the handshake completes successfully.

Refs codership/galera-bugs#957