ASIO patch: reading the event counter when timers are used
We must read the 8-byte timeout expiration counter (using the
"timerfd" descriptor) every time when an EPOLLIN event occurs.
Otherwise, the timer may stop notifying the application of new
events, and all actions related to this timer will be frozen
on some systems; or vice versa, we will get an infinite loop
of notifications due to the constant readiness to read new data.